# api/openapi.yaml
openapi: 3.1.0
info:
  title: LMS Service API
  version: 1.0.0
  description: Phase-1 LMS REST API. Timestamps are UTC and money uses integer minor units.
servers:
  - url: /
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      operationId: register
      tags: [Authentication]
      security: []
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RegisterRequest'}, example: {email: student@example.com, password: a-long-unique-passphrase, display_name: Example Student}}}}
      responses: {'201': {description: Pending account created, content: {application/json: {schema: {$ref: '#/components/schemas/RegistrationResponse'}}}}, '409': {$ref: '#/components/responses/Problem'}, '422': {$ref: '#/components/responses/Problem'}}
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [Authentication]
      security: []
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginRequest'}, example: {email: student@example.com, password: a-long-unique-passphrase}}}}
      responses: {'200': {$ref: '#/components/responses/Tokens'}, '401': {$ref: '#/components/responses/Problem'}, '429': {$ref: '#/components/responses/Problem'}}
  /api/v1/auth/refresh:
    post: {operationId: refresh, tags: [Authentication], security: [], requestBody: {$ref: '#/components/requestBodies/RefreshToken'}, responses: {'200': {$ref: '#/components/responses/Tokens'}, '401': {$ref: '#/components/responses/Problem'}}}
  /api/v1/auth/logout:
    post: {operationId: logout, tags: [Authentication], requestBody: {$ref: '#/components/requestBodies/RefreshToken'}, responses: {'204': {description: Current session revoked}, '401': {$ref: '#/components/responses/Problem'}}}
  /api/v1/auth/logout-all:
    post: {operationId: logoutAll, tags: [Authentication], responses: {'204': {description: All sessions revoked}, '401': {$ref: '#/components/responses/Problem'}}}
  /api/v1/auth/me:
    get: {operationId: currentUser, tags: [Authentication], responses: {'200': {description: Current user, content: {application/json: {schema: {$ref: '#/components/schemas/AuthUser'}}}}, '401': {$ref: '#/components/responses/Problem'}}}
  /api/v1/auth/change-password:
    post: {operationId: changePassword, tags: [Authentication], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChangePasswordRequest'}}}}, responses: {'204': {description: Password changed and sessions revoked}, '422': {$ref: '#/components/responses/Problem'}}}
  /api/v1/auth/forgot-password:
    post: {operationId: forgotPassword, tags: [Authentication], security: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ForgotPasswordRequest'}, example: {email: student@example.com}}}}, responses: {'202': {description: Generic accepted response}, '429': {$ref: '#/components/responses/Problem'}}}
  /api/v1/auth/reset-password:
    post: {operationId: resetPassword, tags: [Authentication], security: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ResetPasswordRequest'}}}}, responses: {'204': {description: Password reset}, '401': {$ref: '#/components/responses/Problem'}}}
  /api/v1/auth/verify-email:
    post: {operationId: verifyEmail, tags: [Authentication], security: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TokenRequest'}}}}, responses: {'204': {description: Email verified}, '401': {$ref: '#/components/responses/Problem'}}}
  /api/v1/users/me:
    patch: {operationId: updateCurrentProfile, tags: [Users], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProfileWrite'}}}}, responses: {'200': {$ref: '#/components/responses/OK'}, '422': {$ref: '#/components/responses/Problem'}}}
  /api/v1/admin/users:
    get: {operationId: adminListUsers, tags: [Users], parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}, {name: q, in: query, schema: {type: string, maxLength: 200}}, {name: status, in: query, schema: {type: string, enum: [pending, active, suspended, disabled]}}, {name: role, in: query, schema: {type: string, enum: [student, teacher, admin]}}], responses: {'200': {$ref: '#/components/responses/List'}, '403': {$ref: '#/components/responses/Problem'}}}
  /api/v1/admin/users/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get: {operationId: adminGetUser, tags: [Users], responses: {'200': {$ref: '#/components/responses/OK'}, '404': {$ref: '#/components/responses/Problem'}}}
  /api/v1/admin/users/{id}/status:
    parameters: [{$ref: '#/components/parameters/ID'}]
    patch: {operationId: adminSetUserStatus, tags: [Users], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserStatusWrite'}, example: {status: suspended}}}}, responses: {'204': {description: Account state changed}, '409': {$ref: '#/components/responses/Problem'}}}
  /api/v1/admin/users/{id}/roles:
    parameters: [{$ref: '#/components/parameters/ID'}]
    put: {operationId: adminReplaceUserRoles, tags: [Users], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RoleWrite'}, example: {roles: [teacher]}}}}, responses: {'200': {$ref: '#/components/responses/OK'}, '409': {$ref: '#/components/responses/Problem'}}}
  /api/v1/courses:
    get: {operationId: listCourses, tags: [Courses], security: [], parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}, {name: q, in: query, schema: {type: string}}, {name: organization_id, in: query, schema: {type: string, format: uuid}}, {name: category_id, in: query, schema: {type: string, format: uuid}}, {name: level, in: query, schema: {type: string, enum: [beginner, intermediate, advanced, all]}}], responses: {'200': {$ref: '#/components/responses/List'}}}
  /api/v1/courses/{slug}:
    parameters: [{$ref: '#/components/parameters/Slug'}]
    get: {operationId: getPublishedCourse, tags: [Courses], security: [], responses: {'200': {description: Course with ordered modules and published lessons}, '404': {$ref: '#/components/responses/Problem'}}}
  /api/v1/teacher/courses:
    get: {operationId: teacherListCourses, tags: [Courses], parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}], responses: {'200': {$ref: '#/components/responses/List'}}}
    post: {operationId: teacherCreateCourse, tags: [Courses], requestBody: {$ref: '#/components/requestBodies/CourseWrite'}, responses: {'201': {$ref: '#/components/responses/Course'}, '422': {$ref: '#/components/responses/Problem'}}}
  /api/v1/teacher/courses/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get: {operationId: teacherGetCourse, tags: [Courses], responses: {'200': {description: Course detail}, '403': {$ref: '#/components/responses/Problem'}}}
    patch: {operationId: teacherUpdateCourse, tags: [Courses], requestBody: {$ref: '#/components/requestBodies/CourseWrite'}, responses: {'200': {$ref: '#/components/responses/Course'}, '409': {$ref: '#/components/responses/Problem'}}}
  /api/v1/teacher/courses/{id}/submit-review:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post: {operationId: teacherSubmitCourseReview, tags: [Courses], responses: {'200': {$ref: '#/components/responses/Course'}}}
  /api/v1/admin/courses/{id}/publish:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post: {operationId: adminPublishCourse, tags: [Courses], responses: {'200': {$ref: '#/components/responses/Course'}, '422': {$ref: '#/components/responses/Problem'}}}
  /api/v1/teacher/courses/{id}/modules:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post: {operationId: teacherCreateModule, tags: [Courses], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ModuleWrite'}, example: {title: Foundations, position: 1}}}}, responses: {'201': {$ref: '#/components/responses/OK'}}}
  /api/v1/teacher/modules/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    patch: {operationId: teacherUpdateModule, tags: [Courses], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ModuleWrite'}}}}, responses: {'200': {$ref: '#/components/responses/OK'}}}
    delete: {operationId: teacherDeleteModule, tags: [Courses], responses: {'204': {description: Draft module deleted}, '409': {$ref: '#/components/responses/Problem'}}}
  /api/v1/teacher/modules/{id}/lessons:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post: {operationId: teacherCreateLesson, tags: [Courses], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LessonWrite'}}}}, responses: {'201': {$ref: '#/components/responses/OK'}}}
  /api/v1/teacher/lessons/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    patch: {operationId: teacherUpdateLesson, tags: [Courses], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LessonWrite'}}}}, responses: {'200': {$ref: '#/components/responses/OK'}}}
    delete: {operationId: teacherDeleteLesson, tags: [Courses], responses: {'204': {description: Lesson deleted}, '409': {$ref: '#/components/responses/Problem'}}}
  /api/v1/media/upload-intents:
    post:
      operationId: createUploadIntent
      tags: [Media]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UploadIntentWrite'}, example: {kind: pdf, filename: handout.pdf, content_type: application/pdf, size_bytes: 24576}}}}
      responses:
        '201': {description: Direct-upload intent, content: {application/json: {schema: {$ref: '#/components/schemas/UploadIntent'}}}}
        '422': {$ref: '#/components/responses/Problem'}
  /api/v1/media/{id}/complete:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post: {operationId: completeMediaUpload, tags: [Media], responses: {'202': {$ref: '#/components/responses/OK'}, '409': {$ref: '#/components/responses/Problem'}}}
  /api/v1/media/{id}/access-url:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get: {operationId: getMediaAccessURL, tags: [Media], responses: {'200': {$ref: '#/components/responses/OK'}, '403': {$ref: '#/components/responses/Problem'}}}
  /api/v1/courses/{id}/enroll:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post: {operationId: enrollFreeCourse, tags: [Learning], responses: {'201': {description: Active enrollment, content: {application/json: {schema: {$ref: '#/components/schemas/Enrollment'}}}}, '402': {$ref: '#/components/responses/Problem'}, '409': {$ref: '#/components/responses/Problem'}}}
  /api/v1/student/enrollments:
    get: {operationId: studentListEnrollments, tags: [Learning], parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}], responses: {'200': {$ref: '#/components/responses/List'}}}
  /api/v1/student/enrollments/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get: {operationId: studentGetEnrollment, tags: [Learning], responses: {'200': {description: Owned enrollment}, '403': {$ref: '#/components/responses/Problem'}}}
  /api/v1/student/enrollments/{id}/lessons/{lessonId}/progress:
    parameters: [{$ref: '#/components/parameters/ID'}, {name: lessonId, in: path, required: true, schema: {type: string, format: uuid}}]
    put: {operationId: studentUpdateProgress, tags: [Learning], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProgressWrite'}}}}, responses: {'200': {description: Idempotently updated progress, content: {application/json: {schema: {$ref: '#/components/schemas/Progress'}}}}, '403': {$ref: '#/components/responses/Problem'}}}
  /api/v1/admin/enrollments:
    get: {operationId: adminListEnrollments, tags: [Learning], parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}, {name: status, in: query, schema: {type: string}}, {name: course_id, in: query, schema: {type: string, format: uuid}}, {name: student_id, in: query, schema: {type: string, format: uuid}}], responses: {'200': {$ref: '#/components/responses/List'}}}
  /api/v1/teacher/quizzes:
    get: {operationId: teacherListQuizzes, tags: [Assessments], parameters: [{$ref: '#/components/parameters/CourseID'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}], responses: {'200': {$ref: '#/components/responses/List'}}}
    post: {operationId: teacherCreateQuiz, tags: [Assessments], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/QuizWrite'}}}}, responses: {'201': {$ref: '#/components/responses/OK'}}}
  /api/v1/teacher/quizzes/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get: {operationId: teacherGetQuiz, tags: [Assessments], responses: {'200': {description: Quiz and answer-key question bank}}}
    patch: {operationId: teacherUpdateQuiz, tags: [Assessments], parameters: [{name: status, in: query, schema: {type: string, enum: [draft, published, archived]}}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/QuizWrite'}}}}, responses: {'200': {$ref: '#/components/responses/OK'}}}
    delete: {operationId: teacherDeleteQuiz, tags: [Assessments], responses: {'204': {description: Draft quiz deleted}, '409': {$ref: '#/components/responses/Problem'}}}
  /api/v1/teacher/quizzes/{id}/questions:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post: {operationId: teacherCreateQuizQuestion, tags: [Assessments], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/QuestionWrite'}}}}, responses: {'201': {$ref: '#/components/responses/OK'}}}
  /api/v1/teacher/quiz-questions/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    patch: {operationId: teacherUpdateQuizQuestion, tags: [Assessments], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/QuestionWrite'}}}}, responses: {'200': {$ref: '#/components/responses/OK'}, '409': {$ref: '#/components/responses/Problem'}}}
    delete: {operationId: teacherDeleteQuizQuestion, tags: [Assessments], responses: {'204': {description: Question deleted}, '409': {$ref: '#/components/responses/Problem'}}}
  /api/v1/student/quizzes/{id}/attempts:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get: {operationId: studentListQuizAttempts, tags: [Assessments], parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}], responses: {'200': {$ref: '#/components/responses/List'}}}
    post: {operationId: studentStartQuizAttempt, tags: [Assessments], responses: {'201': {description: Attempt with no answer-key flags, content: {application/json: {schema: {$ref: '#/components/schemas/QuizAttempt'}}}}, '409': {$ref: '#/components/responses/Problem'}}}
  /api/v1/student/quiz-attempts/{id}/answers:
    parameters: [{$ref: '#/components/parameters/ID'}]
    put: {operationId: studentSaveQuizAnswer, tags: [Assessments], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AnswerWrite'}}}}, responses: {'200': {$ref: '#/components/responses/OK'}, '409': {$ref: '#/components/responses/Problem'}}}
...
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer, bearerFormat: JWT, description: Short-lived LMS access JWT. Refresh tokens are accepted only by refresh/logout.}
...
  responses:
...
    Problem: {description: RFC 7807 problem details, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Tokens: {description: Rotating access and refresh pair, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}
  schemas:
...
    RegisterRequest: {type: object, required: [email, password, display_name], additionalProperties: false, properties: {email: {type: string, format: email, maxLength: 320}, password: {type: string, format: password, minLength: 12, maxLength: 1024, writeOnly: true}, display_name: {type: string, minLength: 2, maxLength: 120}}}
    LoginRequest: {type: object, required: [email, password], additionalProperties: false, properties: {email: {type: string, format: email, maxLength: 320}, password: {type: string, format: password, minLength: 1, maxLength: 1024, writeOnly: true}}}
...
    ForgotPasswordRequest: {type: object, required: [email], additionalProperties: false, properties: {email: {type: string, format: email}}}
...
    RegistrationResponse: {type: object, required: [user, message], properties: {user: {$ref: '#/components/schemas/AuthUser'}, message: {type: string}, verification_token: {type: string, writeOnly: true, description: Development/test only.}}}
...